LayoutTests/accessibility/gtk/title-and-alt-expected.txt


Functional label:   
This tests the exposure of alt and title information.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS image.title is 'AXTitle: Image alt attr'
FAIL image.description should be AXDescription: Image title attr. Was AXDescription: Image alt attr.
FAIL link.title should be AXTitle: Functional label.. Was AXTitle: Functional label.
PASS link.description is 'AXDescription: Link title attr'
FAIL entry.title should be AXTitle: . Was AXTitle: Entry title attr.
PASS entry.description is 'AXDescription: Entry title attr'
PASS button.title is 'AXTitle: Submit'
PASS button.description is 'AXDescription: Submit button title attr'
PASS successfullyParsed is true

TEST COMPLETE
